% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_datetime_format}
\alias{check_datetime_format}
\title{Check a datetime argument is in a suitable format}
\usage{
check_datetime_format(arg_date)
}
\arguments{
\item{arg_date}{The argument to check}
}
\description{
Check if a given argument is of a proper ISO 8601 format e.g.
 `2023-03-02T21:26:10Z`. If not it will give an informative error message.
}
\keyword{internal}
