% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_params.R
\name{check_logical}
\alias{check_logical}
\title{Check a logical argument is in a suitable format}
\usage{
check_logical(arg_lgl)
}
\arguments{
\item{arg_lgl}{The argument to check}
}
\description{
Check a logical argument is in a suitable format
}
