% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tariff_charges.R
\name{get_agile_prices}
\alias{get_agile_prices}
\title{Get Agile future prices}
\usage{
get_agile_prices(
  region = NULL,
  fuel_type = c("electricity", "gas"),
  days_ahead = 2,
  tz = "UTC"
)
}
\arguments{
\item{region}{The GSP region code (e.g. "H" for Southern England).
If not provided, it will attempt to use saved meter details.}

\item{fuel_type}{Type of fuel, either "electricity" or "gas".
Note: Agile is typically only available for electricity.}

\item{days_ahead}{Number of days ahead to fetch prices for (default: 2).}

\item{tz}{a character string that specifies which time zone to parse the date with. The string
must be a time zone that is recognized by the user's OS.}
}
\value{
a \link[tibble:tibble-package]{tibble} of Agile tariff charge data.
}
\description{
A convenience function to get future Agile tariff prices.
This function automatically uses the Agile Octopus product and constructs
tariff codes based on the region.
}
