% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_meter_gsp.R
\name{get_meter_gsp}
\alias{get_meter_gsp}
\title{Get the GSP of a meter-point.}
\usage{
get_meter_gsp(mpan = NULL)
}
\arguments{
\item{mpan}{The electricity meter-point's MPAN. If \code{NULL} (default) it will
try to use the MPAN stored in the \code{OCTOPUSR_MPAN} environment variable.}
}
\value{
a character of the meter-points GSP.
}
\description{
This endpoint can be used to get the GSP
of a given meter-point.
}
