% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tariff_charges.R
\name{get_tariff_charges}
\alias{get_tariff_charges}
\title{List tariff charges for a product}
\usage{
get_tariff_charges(
  product_code,
  tariff_code,
  fuel_type = c("electricity", "gas"),
  rate_type = NULL,
  period_from = NULL,
  period_to = NULL,
  tz = NULL,
  order_by = c("-period", "period")
)
}
\arguments{
\item{product_code}{The product code for the tariff.}

\item{tariff_code}{The tariff code.}

\item{fuel_type}{Type of fuel, either "electricity" or "gas".}

\item{rate_type}{Type of rate to retrieve. For electricity: "standard-unit-rates",
"day-unit-rates", "night-unit-rates". For gas: "standard-unit-rates".}

\item{period_from}{Show charges from the given datetime (inclusive).
This parameter can be provided on its own.}

\item{period_to}{Show charges to the given datetime (exclusive).
This parameter also requires providing the \code{period_from} parameter
to create a range.}

\item{tz}{a character string that specifies which time zone to parse the date with. The string
must be a time zone that is recognized by the user's OS.}

\item{order_by}{Ordering of results returned. Default is that results are
returned in reverse order from latest available figure.
Valid values:
\itemize{
\item \code{period}, to give results ordered forward.
\item \code{-period}, (default), to give results ordered from most recent backwards.
}}
}
\value{
a \link[tibble:tibble-package]{tibble} of the requested tariff charge data.
}
\description{
Return a list of tariff charges for half-hour periods for a
given product and tariff. This is particularly useful for agile pricing
and other time-of-use tariffs.

Unit of measurement:
\itemize{
\item Electricity tariffs: pence per kWh
\item Gas tariffs: pence per kWh
}
\subsection{Parsing dates}{

To return dates properly parsed \link[lubridate:lubridate-package]{lubridate} is
required. Use the \code{tz} parameter to specify a time zone e.g. \code{tz = "UTC"},
the default (\code{tz = NULL}) will return the dates unparsed, as characters.
}
}
