% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octopus_api.R
\name{octopus_api}
\alias{octopus_api}
\title{Main Octopus API}
\usage{
octopus_api(path, query = NULL, api_key = NULL)
}
\arguments{
\item{path}{the API endpoint}

\item{query}{optional list passed to \code{\link[httr:modify_url]{httr::modify_url()}}}

\item{api_key}{The API key to use for authentication
(not required on all endpoints)}
}
\value{
An Octopus API object
}
\description{
Main Octopus API
}
