% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the Octopus API key}
\usage{
set_api_key(api_key = NULL)
}
\arguments{
\item{api_key}{Your API key. If you are an Octopus Energy customer,
you can generate an API key on the
[developer dashboard](https://octopus.energy/dashboard/developer/).}
}
\value{
No return value, called for side effects.
}
\description{
Set the Octopus API key to use. This will be stored as an
environment variable. You should add `OCTOPUSR_API_KEY = <api_key>` to your
`.Renviron` otherwise you will have to call this function every session.
}
