% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_datetime_format.R
\name{check_datetime_format}
\alias{check_datetime_format}
\title{Check the datetime format}
\usage{
check_datetime_format(date, param_name)
}
\arguments{
\item{date}{a character to check}

\item{param_name}{Parameter name, only used in the error message.}
}
\value{
the \code{date} provided
}
\description{
Check that the provided is in the correct
\href{https://en.wikipedia.org/wiki/ISO_8601}{ISO_8601} format required for the
Octopus API.
}
