% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_consumption.R
\name{get_consumption}
\alias{get_consumption}
\title{Get consumption data}
\usage{
get_consumption(
  mpan,
  serial_number,
  api_key = get_api_key(),
  period_from = NULL,
  period_to = NULL,
  page_size = 100L,
  order_by = c("-period", "period"),
  group_by = c("hour", "day", "week", "month", "quarter")
)
}
\arguments{
\item{api_key}{The API key to use for authentication
(not required on all endpoints)}

\item{group_by}{}
}
\description{
Get consumption data
}
