% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octopus_api.R
\name{octopus_api}
\alias{octopus_api}
\title{Main Octopus API}
\usage{
octopus_api(path, query = NULL, api_key = NULL, use_api_key = FALSE)
}
\arguments{
\item{path}{the API endpoint}

\item{query}{optional list of}

\item{api_key}{Your API key. If you are an Octopus Energy customer,
you can generate an API key from your
\href{https://octopus.energy/dashboard/developer/}{online dashboard}.}

\item{use_api_key}{Should the request use the api_key for authentication}
}
\value{
An Octopus API object
}
\description{
Main Octopus API
}
