% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/octopus_api.R
\name{octopus_api}
\alias{octopus_api}
\title{Main Octopus API}
\usage{
octopus_api(path, query = NULL, api_key = NULL)
}
\arguments{
\item{path}{the API endpoint}
}
\value{
an Octopus API object
}
\description{
Main Octopus API
}
