% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the Octopus API key}
\usage{
set_api_key(api_key = NULL)
}
\arguments{
\item{api_key}{Your API key. If you are an Octopus Energy customer,
you can generate an API key from your
\href{https://octopus.energy/dashboard/developer/}{online dashboard}.}
}
\description{
You can find your API key on the
\href{https://octopus.energy/dashboard/developer/}{Octopus Energy developer dashboard}
}
