% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_key.R
\name{set_api_key}
\alias{set_api_key}
\title{Set the Octopus API key}
\usage{
set_api_key(key = NULL)
}
\arguments{
\item{key}{the Octopus API key to use}
}
\description{
You can find your API key on the \href{https://octopus.energy/dashboard/developer/}{octopus energy developer dashboard}
}
