% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meter_details.R
\name{set_meter_details}
\alias{set_meter_details}
\title{Set the details for your gas/electricity meter}
\usage{
set_meter_details(
  meter_type = c("electricity", "gas"),
  mpan_mprn = NULL,
  serial_number = NULL
)
}
\arguments{
\item{meter_type}{Type of meter-point, electricity or gas}

\item{mpan_mprn}{The electricity meter-point's MPAN or gas meter-point’s MPRN.}

\item{serial_number}{The meter's serial number.}
}
\description{
You can find your meter details on the
\href{https://octopus.energy/dashboard/developer/}{octopus energy developer dashboard}
}
